Add files via upload
Stimulation of Variable Length Subnet Masking (VLSM) is a technique used in IP addressing that allows a network to be divided into subnets of different sizes. This flexibility enables more efficient use of IP addresses by tailoring the subnet size to the specific needs of different segments within a network.